Add files via upload
In this excel sheet you can find all texts from 1999-2019 categorised according to diseases. 
The red -marked texts provided no negations, even though there is the possibility for symptoms to exist.
The rest of the texts, either they gave some negated symptoms, or they weren't used due to appropriateness, t.e the category biochemisty was not annotated.
For that please, check the document with the extracted negated sentences per category.